Add UiState class for token info screen
