+++ app/src/main/res/values/strings.xml
@@ -949,13 +949,13 @@
     <string name="ManageCoins_Search">Name, code or contract address</string>
     <string name="ManageCoins_AddToken">Add Token</string>
     <string name="ManageCoins_NoResults">No results found. Try to add token manually if you know its smart contract address.</string>
-    <string name="ManageCoins_BipsDescription">There are 3 common address formats Bitcoin wallets can use to receive incoming payments:
+    <string name="ManageCoins_BipsDescription">There are 3 common address formats %s wallets can use to receive incoming payments:
 \n
 \n- BIP-44 (older)
 \n- BIP-49 (newer)
 \n- BIP-84 (newest)
 \n
-\nWhile Unstoppable wallet supports all 3 of them it recommends to use a Bitcoin wallet operating in BIP-84 format. The wallets operating in this format get to pay lowest fees when sending Bitcoin transactions.</string>
+\nWhile Unstoppable wallet supports all 3 of them it recommends to use a %s wallet operating in BIP-84 format. The wallets operating in this format get to pay lowest fees when sending %s transactions.</string>
     <string name="ManageCoins_BchTypeDescription">There are 2 address formats Bitcoin Cash wallets can use to receive incoming payments:
 \n
 \n- TYPE0 (older)
